Update README with new commands
