+++ Core/RequestHandler.cs
@@ -23,8 +23,7 @@ public class RequestHandler : IExternalEventHandler
         { "PlaceViewsOnSheet", new PlaceViewsOnSheetCommand() },
         { "NewSharedParameter", new NewSharedParameterCommand() },
         { "SetParameters", new SetParametersCommand() },
-        { "GetParameters", new GetParametersCommand() },
-        { "GetParametersByID", new GetParametersByIdCommand() },
+        { "GetElementParameters", new GetElementParametersCommand() },
         { "FilterByParameter", new FilterByParameterCommand() },
         { "ListElementsByCategory", new ListElementsCommand() },
         { "ExecutePlan", new PlanExecutorCommand() },
